BG flag rules Fix small fix to conversion warnings http://arcemu.org/forums/index.php?showtopic=843&st=0&gopid=26960&#entry26960

+++ src/arcemu-world/ObjectMgr.h
@@ -687,7 +687,7 @@ public:
 			++it;
 		}
 		if ( Entries.empty() ) return 0;
-		uint32 n = (Entries.size()>1 ? RandomUInt(  Entries.size()-1  ) : 0);
+		uint32 n = (Entries.size()>1 ? RandomUInt(  ((uint32)Entries.size())-1  ) : 0);
 		return Entries[n];
 	}
